w32: Add cast for assuan_fd_t.
* src/assuan-support.c (my_pipe, my_close, my_read, my_write, my_spawn)
(my_socketpair, my_connect): Cast assuan_fd_t between int.
* src/engine-assuan.c (start): Likewise.
* src/engine-g13.c (start): Likewise.
* src/engine-gpgsm.c (gpgsm_new, prepare): Likewise.

--

+++ src/engine-assuan.c
@@ -716,7 +716,7 @@ start (engine_llass_t llass, const char *command)
     return gpg_error (GPG_ERR_GENERAL);	/* FIXME */
   /* For now... */
   for (i = 0; i < nfds; i++)
-    fdlist[i] = (int) afdlist[i];
+    fdlist[i] = (int)(intptr_t) afdlist[i];
 
   /* We "duplicate" the file descriptor, so we can close it here (we
      can't close fdlist[0], as that is closed by libassuan, and
